refactored files into separate directories; fixed a bug
